Move scoring out of general Util into scoring util file
